+++ lkinpractice/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
+from lkinpractice.models import Lkin, LkinBusiness, GenericProfile
 
-# Register your models here.
+
+admin.site.register([Lkin, LkinBusiness, GenericProfile])