+++ lkinpractice/views.py
@@ -1,10 +1,9 @@
 from django.shortcuts import redirect,render
 from lkinpractice.forms import LkinForm
-from lkinpractice.models import Lkin
+from lkinpractice.models import Lkin, GenericProfile
+from django.views.generic import View, TemplateView
+from django.db.models import Q
 
-# Create your views here.
-
-from django.views.generic import View
 
 
 class LkinView(View):
@@ -22,6 +21,14 @@ class LkinView(View):
         lkins = Lkin.objects.all()
         return render(request, 'lkinpractice/lkin_practice.html', {'form': form, 'lkins': lkins})
         
-# class ProfileView()
         
+class LkinProfile(TemplateView):
+    template_name = "lkinpractice/lkin_profile.html"
+    
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["profile"] = GenericProfile.objects.filter(Q(lkin__user=self.request.user) | Q(lkinbusiness__user=self.request.user)
+            )[0]
+        return context
+