+++ settings/urls.py
@@ -7,6 +7,8 @@ from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
 
+handler404 = 'home.views.error_404'
+
 urlpatterns = [
     path('', include('home.urls')),
     path('polls/', include('polls.urls')),